ingestion-service/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12  # required by FastAPI for multipart/form-data uploads

# JWT validation against Keycloak's JWKs endpoint — same lib + version
# as langgraph-service so behavior matches.
python-jose[cryptography]==3.3.0

# httpx for both Keycloak JWKs fetch and (in Phase 3) the embedding
# model + Qdrant calls. Pinned to match langgraph-service.
httpx==0.27.2

pydantic==2.9.2

# OTel — same versions as langgraph-service post-langfuse-v3 bump so
# the API trio stays aligned across services. Range-pinned until first
# successful build resolves an exact version.
opentelemetry-sdk>=1.33.1,<2.0.0
opentelemetry-exporter-otlp-proto-grpc>=1.33.1,<2.0.0
opentelemetry-instrumentation-fastapi>=0.54b0,<1.0.0
opentelemetry-instrumentation-httpx>=0.54b0,<1.0.0
opentelemetry-instrumentation-logging>=0.54b0,<1.0.0

# Phase 3 deps (commented for now; uncomment + push when ingestion logic lands):
# unstructured[pdf,docx]==0.16.5    # document parser
# langchain-text-splitters==0.3.2   # RecursiveCharacterTextSplitter
# qdrant-client==1.12.1             # Qdrant write client
# langchain-openai==0.2.14          # OpenAI-compat embeddings client (vLLM)
